--- src/components/List/commonPropsKeys.tsx
@@ -1,13 +1,0 @@
-import { ListCommonProps } from './ListItem'
-
-type ListCommonProp = keyof ListCommonProps
-
-const commonPropsKeys: ListCommonProp[] = [
-	'color',
-	'size',
-	'icon',
-	'colorTheme',
-	'bulletColor'
-]
-
-export default commonPropsKeys
