+++ src/components/List/index.tsx
@@ -4,8 +4,7 @@ import React, { HTMLAttributes } from 'react'
 import { GenericComponentProps } from '../../interfaces/GenericComponentProps'
 import { MarginProps } from '../common/Spacing/MarginProps'
 import { PaddingProps } from '../common/Spacing/PaddingProps'
-import commonPropsKeys from './commonPropsKeys'
-import { ListCommonProps } from './ListItem'
+import { commonPropsKeys, ListCommonProps } from './ListCommonProps'
 import { StyledList } from './ListStyles'
 
 export interface ListProps
@@ -20,15 +19,11 @@ export const List: React.FC<ListProps> = ({ children, ...props }) => {
 
 	return (
 		<StyledList {...props}>
-			{React.Children.map(children, (item) => {
-				if (React.isValidElement(item)) {
-					return React.cloneElement(
-						item,
-						Object.assign({}, commonProps, item.props)
-					)
-				}
-				return item
-			})}
+			{React.Children.map(children, (item) =>
+				React.isValidElement(item)
+					? React.cloneElement(item, Object.assign({}, commonProps, item.props))
+					: item
+			)}
 		</StyledList>
 	)
 }
